// src/types.ts
export type Planet = {
  geology: {
    content: string;
    source: string;
  };
  images: {
    geology: string;
    internal: string;
    planet: string;
  };
  name: string;
  overview: {
    content: string;
    source: string;
  };
  radius: string;
  revolution: string;
  rotation: string;
  structure: {
    content: string;
    source: string;
  };
  temperature: string;
};